Moved the Authentication code in user module
