+++ CMMSAPIs/BS/Users/UserAccessBS.cs
@@ -1,7 +1,9 @@
 ﻿using CMMSAPIs.Helper;
+using CMMSAPIs.Models.Authentication;
 using CMMSAPIs.Models.Users;
 using CMMSAPIs.Models.Utils;
 using CMMSAPIs.Repositories.Users;
+using Microsoft.Extensions.Configuration;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
@@ -10,7 +12,7 @@ namespace CMMSAPIs.BS.Users
 {
     public interface IUserAccessBS
     {
-
+        Task<UserToken> Authenticate(CMUserCrentials userCrentials);
         public Task<CMUserDetail> GetUserDetail(int user_id);
         public Task<List<CMUser>> GetUserList(int facility_id);
         public Task<CMDefaultResponse> CreateUser(CMUserDetail request);
@@ -25,13 +27,30 @@ namespace CMMSAPIs.BS.Users
 
     public class UserAccessBS : IUserAccessBS
     {
+        private readonly IConfiguration _configuration;
         private readonly DatabaseProvider databaseProvider;
         private MYSQLDBHelper getDB => databaseProvider.SqlInstance();
-        public UserAccessBS(DatabaseProvider dbProvider)
+        public UserAccessBS(DatabaseProvider dbProvider, IConfiguration configuration)
         {
+            _configuration   = configuration;
             databaseProvider = dbProvider;
         }
 
+        public async Task<UserToken> Authenticate(CMUserCrentials userCrentials)
+        {
+            try
+            {
+                using (var repos = new UserAccessRepository(getDB, _configuration))
+                {
+                    return await repos.Authenticate(userCrentials);
+                }
+
+            }
+            catch (Exception ex)
+            {
+                throw;
+            }
+        }
         public async Task<CMUserDetail> GetUserDetail(int user_id)
         {
             try
@@ -40,7 +59,7 @@ namespace CMMSAPIs.BS.Users
                 {
                     return await repos.GetUserDetail(user_id);
                 }
-            }
+            }   
             catch (Exception ex)
             {
                 throw;
